게시물 삭제 시 작성자만 지울 수 있게 수정

+++ src/main/java/com/lhw/thymeleaf/jpaBoard/article/controller/ArticleController.java
@@ -142,8 +142,20 @@ public class ArticleController {
 
     @RequestMapping("/doDelete")
     @ResponseBody
-    public String doDelete(long id) {
-        Article article = articleRepository.findById(id).get();
+    public String doDelete(long id, HttpSession session) {
+        boolean isLogined = false;
+        long loginedUserId = 0;
+        if (session.getAttribute("loginedUserId") != null) {
+            isLogined = true;
+            loginedUserId = (long)session.getAttribute("loginedUserId");
+        } else {
+            return """
+                <script>
+                alert('로그인을 해주세요!');
+                history.back();
+                </script>
+                """;
+        }
 
         if (!articleRepository.existsById(id)) {
             return """
@@ -152,6 +164,16 @@ public class ArticleController {
                     history.back();
                     </script>
                     """.formatted(id);
+        }
+        Article article = articleRepository.findById(id).get();
+
+        if (article.getUser().getId() != loginedUserId) {
+            return """
+                    <script>
+                    alert('권한이 없습니다.');
+                    history.back();
+                    </script>
+                    """;
         }
         articleRepository.deleteById(id);
         return """
